requirements.txt
streamlit==1.39.0
pypdf==4.3.1
reportlab==4.2.2
Pillow==10.4.0
